check for existence of documents first before creating events and userEvent objects in controllers

+++ server/controllers/eventController.js
@@ -3,6 +3,11 @@ const db = require('../models');
 
 module.exports = {
 
+  /** 
+   * @param {Request} req 
+   * @param {Response} res 
+   * @returns {Array<User>}
+   */
   findAll: function (req, res) {
     db.Event
       .find({})
@@ -20,8 +25,13 @@ module.exports = {
 
   create: async function (req, res) {
     try {
+      const user = await db.User.findById(req.body.creator); // Check if user exists
+      if (!user) {
+        return res.status(404).json({ message: `User with id ${req.body.creator} does not exist.` }); 
+      }
+      
       const event = await db.Event.create(req.body);
-      await db.UserEvent.create({user_id: event.creator, event_id: event._id}); // Event creator will attend event
+      await db.UserEvent.create({ user_id: event.creator, event_id: event._id }); // Event creator will attend event
       return res.json(event);
     } catch (err) {
       return res.status(422).json(err);
@@ -31,7 +41,12 @@ module.exports = {
   update: function (req, res) {
     db.Event
       .findOneAndUpdate({ _id: req.params.id }, req.body)
-      .then(dbModel => res.json(dbModel))
+      .then(dbModel => {
+        if (!dbModel) { // Check if event existed
+          return res.status(404).json({ message: `Event with id ${req.params.id} does not exist.` })
+        }
+        return res.json(dbModel)
+      })
       .catch(err => res.status(422).json(err));
   },
 
@@ -39,7 +54,10 @@ module.exports = {
     db.Event
       .findByIdAndDelete(req.params.id)
       .then(async dbModel => {
-        await db.UserEvent.deleteMany({event_id: dbModel._id}); // Delete all UserEvent documents for the deleted event
+        if (!dbModel) { // Check if event existed
+          return res.status(404).json({ message: `Event with id ${req.params.id} does not exist.` })
+        }
+        await db.UserEvent.deleteMany({ event_id: dbModel._id }); // Delete all UserEvent documents for the deleted event
         return res.json(dbModel);
       })
       .catch(err => res.status(422).json(err));
